Moved class Hal out of /vent/io/__init__.py and into /vent/io/hal.py
